Lets hope this doesnt break everything
Im not sure why there are so many conflicts now. I added some functionality with userIDs and made the cart. Thing should be working, but there are also a bunch of dupe files. I will be cleaning that up tomorrow but if you want to check anything make sure you are not running the files in project root.